Applied patches 1993189, 1993256, 1993297, 1993302, 1993317, 1993335, and 1993359 from sourceforge, submitted by Jerry James. The purpose of the patches is to clean up a large number of compiler warnings related to generics and type casts.
git-svn-id: https://findbugs.googlecode.com/svn/trunk@9738 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3

+++ findbugs/src/java/edu/umd/cs/findbugs/ba/BytecodeScanner.java
@@ -56,7 +56,7 @@ public class BytecodeScanner implements org.apache.bcel.Constants {
 			result = (short) (value & 0x7F);
 			result |= 0x80;
 		} else {
-			result = (short) value;
+			result = value;
 		}
 		return result;
 	}
