+++ findbugs/src/java/edu/umd/cs/findbugs/ba/vna/ValueNumberDataflow.java
@@ -62,7 +62,7 @@ public class ValueNumberDataflow extends AbstractDataflow<ValueNumberFrame, Valu
 		int shift = isStatic ? 0 : 1;
 		for (int i = 0; i < numParams; ++i) {
 			valueNumberToParamMap.put(
-					frameAtEntry.getValue(i + shift), (Integer)(i));
+					frameAtEntry.getValue(i + shift), i);
 		}
 
 		return valueNumberToParamMap;
